requirements.txt
aioamqp==0.13.0
appnope==0.1.0
argh==0.26.2
astroid==2.2.5
asyncpg==0.18.3
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
backports.tempfile==1.0
backports.weakref==1.0.post1
backtrace==0.2.1
bleach==3.1.0
boto3==1.9.158
botocore==1.12.158
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
colander==0.9.9
colorama==0.3.7
colorful==0.5.0
cycler==0.10.0
Cython==0.29.14
dask==1.2.2
decorator==4.4.0
defusedxml==0.6.0
distributed==1.28.1
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.14
entrypoints==0.3
Flask==0.10.1
future==0.17.1
gender-guesser==0.4.0
gitdb2==2.0.5
GitPython==2.1.11
gql==0.1.0
graphql-core==2.2
HeapDict==1.0.0
hyperopt==0.1.2
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
iso8601==0.1.8
isort==4.3.20
itsdangerous==1.1.0
jedi==0.13.3
jenkinsapi==0.3.11
-e git+https://github.com/perrygeo/jenks.git@80b2557d635ba4a0640157d3616b50dbf39c2221#egg=jenks
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
jupyterlab==0.35.6
jupyterlab-server==0.2.0
kiwisolver==1.1.0
lark-parser==0.7.1
lazy-object-proxy==1.4.1
llvmlite==0.29.0
locket==0.2.0
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
msgpack==0.6.1
nbconvert==5.5.0
nbformat==4.4.0
nest-asyncio==1.0.0
networkx==2.3
notebook==5.7.8
numba==0.44.0
numpy==1.16.4
nvidia-ml-py3==7.352.0
packaging==19.1
pamqp==2.3.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
partd==0.3.10
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.0
prometheus-client==0.7.0
promise==2.2.1
prompt-toolkit==2.0.9
psutil==5.6.3
ptvsd==4.2.10
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
pyjson5==0.4.6
pylint==2.3.1
pymongo==3.8.0
pyparsing==2.4.0
pyrsistent==0.15.2
pytest==5.1.2
python-dateutil==2.8.0
python-dotenv==0.10.2
pytz==2019.1
PyYAML==5.1.1
pyzmq==18.0.1
redis==2.8.0
requests==2.22.0
Rx==1.6.1
s3transfer==0.2.0
scikit-learn==0.21.2
scipy==1.3.0
Send2Trash==1.5.0
sentry-sdk==0.10.0
shap==0.32.1
shortuuid==0.5.0
six==1.12.0
sklearn==0.0
smmap2==2.0.5
sortedcontainers==2.1.0
subprocess32==3.5.4
swifter==0.288
tblib==1.4.0
terminado==0.8.2
testpath==0.4.2
toolz==0.9.0
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
translationstring==1.3
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.6.0
wandb==0.8.3
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.2
xgboost==0.90
zict==0.1.4
zipp==0.6.0
